separate kicad pcm proxy to separate file
